Change operator result type from double to int
The idea for the result of the operation to be of double type was to
unify the result for all operators, including division, which returns
non-integer results. However, since the division operator implementation
is actually integer division, the result will also always be an int,
making the floating point precision of double useless.

+++ include/calculator.h
@@ -7,9 +7,9 @@
 #define OPERATION_FAILURE -1
 #define OPERATION_COUNT 4
 
-int addition(int number1, int number2, double * sum);
-int subtraction(int number1, int number2, double * difference);
-int multiplication(int number1, int number2, double * product);
-int division(int number1, int number2, double * quotient);
+int addition(int number1, int number2, int * sum);
+int subtraction(int number1, int number2, int * difference);
+int multiplication(int number1, int number2, int * product);
+int division(int number1, int number2, int * quotient);
 
 #endif
