+++ src/calculator/multiplication.c
@@ -1,7 +1,7 @@
 #include "../../include/calculator.h"
 
-int multiplication(int number1, int number2, double * product){
-	double temp_product = number1 * number2;
+int multiplication(int number1, int number2, int * product){
+	int temp_product = number1 * number2;
 	/* An honest attempt at catching an overflow / underflow */
 	if (number1 != 0 && temp_product/ number1 != number2)
 		return OPERATION_FAILURE;
