+++ src/operation/create_operation.c
@@ -1,6 +1,6 @@
 #include "../../include/operation.h"
 
-Operation * create_operation(char const * name, char symbol,int (* pointer) (int,int,double*)){
+Operation * create_operation(char const * name, char symbol,int (* pointer) (int,int,int*)){
 	Operation * operation = malloc(sizeof(Operation));
 	strncpy(operation->name, name, OPERATION_NAME_SIZE);
 	operation->symbol = symbol;
